Added several features
Now when it guesses a word, it will check to see which letters are in the correct spot, and which are correct but in the wrong spot, and create a string such like "gyy-g".

G means that letter is green, y meaning it was yellow, and a dash representing the greyed out letters that are incorrect.

It will then use that string to loop through a list of all possible correct wordle words, and return a list of the possible words it might be that share similar letters.